refactor: move logic into vaporwave module
